requirements.txt
flask
openai
google-generativeai
gunicorn